refactor: `Grenades`
- 將 `Grenade` 和 `GrenadeHandlers` 按照功能性分離
- 修改資料夾的名稱(`equipment`->`grenade`)
